python/pyproject.toml
[project]
name = "kafka-kraft-demo-python"
version = "0.1.0"
description = "Kafka kraft demo in Python"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.12.2",
    "fastapi>=0.121.3",
    "pydantic>=2.12.4",
    "pyyaml>=6.0.3",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
